Move ValueMerger inside simplify module
